pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "customer-support-rag"
version = "0.1.0"
description = "AI-Powered Telecom Customer Support Assistant using RAG"
authors = [
    {name = "Abhishek Roy", email = "abhishek.roy@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-openai>=0.0.5",
    "openai>=1.12.0",
    "chromadb>=0.4.22",
    "streamlit>=1.31.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.2",
    "pandas>=2.2.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
